fix: move DEFAULT_BLOCK_GAS_LIMIT to constants.ts
