+++ indexer/src/types/header.ts
@@ -10,16 +10,7 @@ import { JsonRpcBlock } from "./types.ts";
 import { KAKAROT } from "../provider.ts";
 
 // Constant
-import { NULL_HASH } from "../constants.ts";
-
-// A default block gas limit in case the call to get_block_gas_limit fails.
-export const DEFAULT_BLOCK_GAS_LIMIT = (() => {
-  const defaultBlockGasLimitStr = Deno.env.get("DEFAULT_BLOCK_GAS_LIMIT");
-  if (!defaultBlockGasLimitStr) {
-    throw new Error("ENV: DEFAULT_BLOCK_GAS_LIMIT is not set");
-  }
-  return defaultBlockGasLimitStr;
-})();
+import { DEFAULT_BLOCK_GAS_LIMIT, NULL_HASH } from "../constants.ts";
 
 /**
  * Converts a Starknet block header to an Ethereum block header in JSON RPC format.
