+++ indexer/tests/header.test.ts
@@ -6,9 +6,9 @@ import {
   PrefixedHexString,
 } from "../src/deps.ts";
 import { assertEquals } from "https://deno.land/std@0.213.0/assert/assert_equals.ts";
-import { DEFAULT_BLOCK_GAS_LIMIT, toEthHeader } from "../src/types/header.ts";
-import { JsonRpcBlock } from "../src/types/types.ts";
-import { padString } from "../src/utils/hex.ts";
+import { JsonRpcBlock, toEthHeader } from "./header.ts";
+import { DEFAULT_BLOCK_GAS_LIMIT } from "../constants.ts";
+import { padString } from "../utils/hex.ts";
 import sinon from "npm:sinon";
 import { KAKAROT } from "../src/provider.ts";
 import { NULL_HASH } from "../src/constants.ts";
